+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The `proof` field in broadcasted invoke v3 transactions is now a base64-encoded byte blob (`Vec<u8>`) instead of base64-encoded packed `u32` values (`Vec<u32>`). This reflects the upstream change to use compressed proofs.
 - The `blockifier` crate has been upgraded to 0.18.0-rc.1 (and its dependencies to match that).
+- HTTP request bodies for `add_invoke_transaction` are compressed with `gzip` if the transaction contains a nonempty proof. Compression can be disabled in custom networks using the `gateway.compress-requests` CLI flag.
 
 ## [0.22.0] - 2026-03-19